moved sorting logic into items in utils
