+++ tests/item.test.js
@@ -1,0 +1,12 @@
+/**
+ * placeholders
+ *
+ * import from utils
+ *
+ *
+ * test data should include
+ *
+ * names with characters/symbols to start
+ * return message based on boolean
+ * arrays of pairs of items
+ */
